added curly brackets to isValidDefaultValue, moved validation methods into schema, reversed conditional statement for validation warning
