+++ src/core/schema.js
@@ -1,7 +1,6 @@
 var debug = require('../utils/debug');
 var propertyTypes = require('./propertyTypes').propertyTypes;
 var warn = debug('core:schema:warn');
-var isValidDefaultValue = require('../utils/validate-schema').isValidDefaultValue;
 
 /**
  * A schema is classified as a schema for a single property if:
@@ -73,14 +72,15 @@ function processPropertyDefinition (propDefinition) {
   // Fill in type name.
   propDefinition.type = typeName;
 
-  // Fill in default value.
-  if (!('default' in propDefinition)) {
-    propDefinition.default = propType.default;
-  } else {
-    // check if valid default value
+  // Check if default value exists.
+  if ('default' in propDefinition) {
+    // Check if default values are valid
     if (!isValidDefaultValue(typeName, defaultVal)) {
       warn('The default value: ' + '"' + defaultVal + '"' + ' does not match the type: ' + typeName);
     }
+  } else {
+    // Fill in default value.
+    propDefinition.default = propType.default;
   }
 
   return propDefinition;
@@ -166,3 +166,65 @@ function stringifyProperty (value, propDefinition) {
   return propDefinition.stringify(value);
 }
 module.exports.stringifyProperty = stringifyProperty;
+
+/**
+ * Checks if Valid default coordinates
+ * @param {unknown} possibleCoordinates
+ * @param {number} dimensions - 2 for 2D Vector or 3 for 3D vector
+ * @returns {boolean} A boolean determining if coordinates are parsed correctly.
+ */
+function isValidDefaultCoordinate (possibleCoordinates, dimensions) {
+  if (typeof possibleCoordinates !== 'object' || possibleCoordinates === null) {
+    return false;
+  } else if (Object.keys(possibleCoordinates).length !== dimensions) {
+    return false;
+  } else {
+    if (dimensions === 2 && (possibleCoordinates.x === 0 || possibleCoordinates.x) && (possibleCoordinates.y === 0 || possibleCoordinates.y)) {
+      if (typeof possibleCoordinates.x === 'number' && typeof possibleCoordinates.y === 'number') {
+        return true;
+      }
+    }
+    if (dimensions === 3 && (possibleCoordinates.x === 0 || possibleCoordinates.x) && (possibleCoordinates.y === 0 || possibleCoordinates.y) && (possibleCoordinates.z === 0 || possibleCoordinates.z)) {
+      if (typeof possibleCoordinates.x === 'number' && typeof possibleCoordinates.y === 'number' && typeof possibleCoordinates.z === 'number') {
+        return true;
+      }
+    }
+    if (dimensions === 4 && (possibleCoordinates.x === 0 || possibleCoordinates.x) && (possibleCoordinates.y === 0 || possibleCoordinates.y) && (possibleCoordinates.z === 0 || possibleCoordinates.z) && (possibleCoordinates.w === 0 || possibleCoordinates.w)) {
+      if (typeof possibleCoordinates.x === 'number' && typeof possibleCoordinates.y === 'number' && typeof possibleCoordinates.z === 'number' && typeof possibleCoordinates.w === 'number') {
+        return true;
+      }
+    }
+  }
+
+  return false;
+}
+module.exports.isValidDefaultCoordinate = isValidDefaultCoordinate;
+
+/**
+ * Validates the default values in a schema
+ * @param {string} type - type in a prop
+ * @param {unknown} defaultVal - default in a prop
+ * @returns {boolean} A boolean determining if defaults are valid.
+ */
+function isValidDefaultValue (type, defaultVal) {
+  if (type === 'audio' && typeof defaultVal !== 'string') { return false; }
+  if (type === 'array' && !Array.isArray(defaultVal)) { return false; }
+  if (type === 'asset' && typeof defaultVal !== 'string') { return false; }
+  if (type === 'boolean' && typeof defaultVal !== 'boolean') { return false; }
+  if (type === 'color' && typeof defaultVal !== 'string') { return false; }
+  if (type === 'int' && typeof defaultVal !== 'number') { return false; }
+  if (type === 'number' && typeof defaultVal !== 'number') { return false; }
+  if (type === 'map' && typeof defaultVal !== 'string') { return false; }
+  if (type === 'model' && typeof defaultVal !== 'string') { return false; }
+  if (type === 'selector' && typeof defaultVal !== 'string') { return false; }
+  if (type === 'selectorAll' && typeof defaultVal !== 'string') { return false; }
+  if (type === 'src' && typeof defaultVal !== 'string') { return false; }
+  if (type === 'string' && typeof defaultVal !== 'string') { return false; }
+  if (type === 'time' && typeof defaultVal !== 'number') { return false; }
+  if (type === 'vec2') { return isValidDefaultCoordinate(defaultVal, 2); }
+  if (type === 'vec3') { return isValidDefaultCoordinate(defaultVal, 3); }
+  if (type === 'vec4') { return isValidDefaultCoordinate(defaultVal, 4); }
+
+  return true;
+}
+module.exports.isValidDefaultValue = isValidDefaultValue;
